refactor: components 폴더에 ai_report 폴더 생성
