+++ src/components/ai_report/lawyer_card.py
