+++ src/components/ai_report/sidebar.py
